[FEATURE] Introduce PreviewRenderer pattern
This introduces a new approach to registering and rendering
previews; for content elements initially but possible to apply
to any record type, and possible to call from other contexts
than the PageLayoutView, e.g. AJAX-based preview rendering.

Basically, this turns the old hook approach into a proper
pattern where preview renderers are registered for a specific
CType and must implement proper interfaces. A Resolver
pattern is also introduced with a standard implementation and
a standard renderer is registered for backwards compatibility.

Change-Id: Ibf85d9b50b7bc6506d72c1ee63078373eaf9e433