[BUGFIX] Avoid duplicate calls to update reference index
This patch adds a check when processing the update
reference index stack, in order to avoid calling the
update function more than once with the same input
arguments (which are $table and $uid).

Saves around 100 SQL queries when doing a standard
copy of 25 (nested) pages with 25 content elements.

Change-Id: If162218d4eabb749a92270e18c7c39108fd15f34
Resolves: #78761
Releases: master, 7.6